Refactor module interfaces
