+++ lib/interface/src/lib.rs
@@ -1,9 +1,9 @@
 mod declaration;
 mod module;
+mod record_definition;
 mod type_alias;
-mod type_definition;
 
 pub use declaration::*;
 pub use module::*;
+pub use record_definition::*;
 pub use type_alias::*;
-pub use type_definition::*;
